Breaking change - change HOC names, and split them to separate files
