+++ lib/HOC/analytics-context-types.ts
@@ -1,0 +1,3 @@
+import * as PropTypes from 'prop-types';
+
+export default { analytics: PropTypes.object };
