+++ lib/HOC/without-analytics.ts
@@ -1,0 +1,3 @@
+import { mapProps, InferableComponentEnhancerWithProps } from 'recompose';
+
+export const withoutAnalytics = mapProps(({ analytics, ...otherProps }) => otherProps);
